No but like for real this time lol

+++ src/game/save_file.c
@@ -685,7 +685,7 @@ void save_file_set_camera_fields(s8 cameraType, s8 cameraSpeed) {
 
 s8 save_file_get_camera_type(void) {
     struct SaveFile *saveFile = &gSaveBuffer.files[gCurrSaveFileNum - 1][0];
-    if (saveFile->cameraType >= USERCAM_COUNT)
+    if ((u8) saveFile->cameraType >= USERCAM_COUNT)
         return 0;
 
     return saveFile->cameraType;
@@ -693,7 +693,7 @@ s8 save_file_get_camera_type(void) {
 
 s8 save_file_get_camera_speed(void) {
     struct SaveFile *saveFile = &gSaveBuffer.files[gCurrSaveFileNum - 1][0];
-    if (saveFile->cameraSpeed >= NUM_CAMERA_SPEED_OPTIONS)
+    if ((u8) saveFile->cameraSpeed >= NUM_CAMERA_SPEED_OPTIONS)
         return 0;
 
     return saveFile->cameraSpeed;
